feat add small refactor
