+++ src/components/PaginationInvoice/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PaginationInvoice'
