+++ src/pages/InvoicePage/utils/generateInvoicePDF.ts
@@ -8,8 +8,8 @@ export const generatePDF = async (componentRef: RefObject<HTMLDivElement>) => {
   dayjs.extend(localizedFormat)
   const date = dayjs().format('LLLL')
 
-  if (componentRef.current) {
-    const canvas = await html2canvas(componentRef.current, { scale: 3 })
+  if (componentRef?.current) {
+    const canvas = await html2canvas(componentRef?.current, { scale: 3 })
     const imgData = canvas.toDataURL('image/png')
     const pdf = new jsPDF({
       orientation: 'portrait',
